Move Exceptions Together
Move the exception related classes together into their own folder &
namespace.
