+++ src/Aydsko.iRacingData/Exceptions/iRacingDataClientException.cs
@@ -1,7 +1,7 @@
 ﻿// © 2022 Adrian Clark
 // This file is licensed to you under the MIT license.
 
-namespace Aydsko.iRacingData;
+namespace Aydsko.iRacingData.Exceptions;
 
 [Serializable]
 public class iRacingDataClientException : Exception
